requirements.txt
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
constantly==15.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
multidict==3.3.1
Pillow==5.0.0
requests==2.18.4
six==1.11.0
urllib3==1.22
yarl==0.17.0
zope.interface==4.4.3
breadability==0.1.20
certifi==2017.11.5
chardet==3.0.4
docopt==0.6.2
idna==2.6
jieba==0.39
nltk==3.2.5
numpy==1.14.0
sumy==0.7.0
pyhooked==0.8.1
selenium
pypiwin32
baidu-aip==2.0.0.1